+++ src/components/Heading/headingChangelogData.ts
@@ -10,21 +10,12 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
-    date: "2023-10-18",
-    version: "2.1.0",
-    type: "Bug Fix",
-    affects: ["Functionality"],
-    notes: [
-      "Fixes an issue with `backgroundColor` and `foregroundColor` props not prioritizing the passed design token values for the `Hero` component.",
-    ],
-  },
-  {
-    date: "2023-9-28",
-    version: "2.0.0",
+    date: "Prerelease",
+    version: "Prerelease",
     type: "Update",
-    affects: ["Styles"],
+    affects: ["Accessibility"],
     notes: [
-      "Applied Typo2023 styles, including font size, font color, and text link patterns.",
+      'Updated the `aria-roledescription` value to "subtitle" (a more familiar and recognizable term) for the `overline` element.',
     ],
   },
 ];